move scoring to actions
